+++ src/parser/enrichers/item/WandOfMagicMissiles.ts
@@ -6,6 +6,7 @@ export default class WandOfMagicMissiles extends DDBEnricherData {
     return {
       type: DDBEnricherData.ACTIVITY_TYPES.DAMAGE,
       addItemConsume: true,
+      addScalingMode: "amount",
       targetType: "creature",
       data: {
         damage: {
@@ -13,14 +14,16 @@ export default class WandOfMagicMissiles extends DDBEnricherData {
           parts: [DDBEnricherData.basicDamagePart({
             number: 3,
             denomination: 4,
+            bonus: "3",
             types: ["force"],
             scalingMode: "whole",
-            scalingFormula: "1d4 + 1",
+            scalingNumber: 1,
+            scalingFormula: "1",
           })],
         },
         "consumption.scaling": {
           allowed: true,
-          max: "@item.uses.max - @item.uses.spent",
+          max: this.is2014 ? "@item.uses.max - @item.uses.spent" : "min(@item.uses.max - @item.uses.spent, 3)",
         },
         range: {
           value: "120",